ticket:1061 - Implemented option No. 2 for handling client requests for an available server
Summary:
When an OpenDb request comes in from a client, rasmgr will try to find a suitable server
for that client. In the previous version only one such attempt was made and if it failed the client
the client received an error message. Now, when an OpenDb request comes from the client, rasmgr will
attempt a fixed number of times to find a free server. Between each attempt, the client request will
be put to sleep to allow rasmgr to manage the list of servers.

Test Plan: Unit tests have already been implemented.

Reviewers: dmisev, mdumitru, gmerticariu

Differential Revision: http://codereview.rasdaman.org/D10