ticket:702 Postgis is installed but still not found (accepted by Dimitar: http://codereview.rasdaman.org/D21)
Summary: I have tested with some version of PostGIS (2.0, 2.1.x) with PostgreSQL 9.1 and 8.4. There are a lot of trouble if just search postgis.sql and spatial.ref.sys.sql (in fact, it can be postgis-2.x.x.sql,...). So I've changed to a better idea to let user make sure postgis has been installed correctly (with PostgreSQL >= 9.1), keep the way to enable with PostgreSQL 8.4 (but user has to provide path to these files).

Test Plan: update_petascopedb.sh --with-postgis

Reviewers: dmisev

Reviewed By: dmisev

Differential Revision: http://codereview.rasdaman.org/D21