ticket:1042 - Updated peer management implementation in rasnet to respect the specification of the old protocol.
Summary:
Peer management syntax conforms to the RNP specification. When a request is received from a client, a rasmgr will try to acquire a free server for the client. If one is not found in the local list of rasservers, and no new servers can be started, the rasmgr will contact its peers to ask for a server.
If a peer replies with the details of a server, the information is forwarded to the client.

Test Plan: The change is covered by unit tests in rasmgr_x/test

Reviewers: gmerticariu, mdumitru

Differential Revision: http://codereview.rasdaman.org/D70