ticket:1155 - Implemented functionality in the configuration manager which saves the state of rasmgr to a unique file if the configuration is not explicitly saved through the rascontrol save command.
Summary:
In the old implementation of rasmgr a user could update rasgmr.conf through the rascontrol save command. If a user modified the internal state of rasmgr through rascontrol, but forgot
to explicitly save the configuration and stopped rasmgr, the configuration was saved to a unique file in the $RMANHOME/etc folder. The new implementation of rasmgr has been changed to support this functionality.

Test Plan:
1. start rasdaman, make a modification through rascontrol, stop_rasdaman -- The expected result is that files containing the backed up data are created in $RMANHOME/etc
2. start rasdaman, make a modification through rascontrol, save, stop_rasdaman -- The expected result is that rasmgr.conf and rasmgr.auth are updated and no temporary files are created.

Reviewers: gmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D77