ticket:1150 - rasmgr log output is now sent to rasmgr.(PID).log as was in the old implementation of rasmgr.
Summary: As decided in today's meeting, I have redirected rasmgr's log output to a file in $RMANHOME/log with the name rasmgr.(PID).log where PID is the process id of the rasmgr process.

Test Plan: When rasdaman is started a file with the name rasmgr.(PID).log is created in $RMANHOME/log with the same PID as rasmgr.

Reviewers: dmisev, mdumitru

Differential Revision: http://codereview.rasdaman.org/D84