ticket:1167 Petascope_insertdemo.sh should keep the band names after update
Summary:
After the last update, the default band name was changed from Gray to b1. This makes things confusing,
especially when running the systemtests that make certain assumptions about the contents of your database.

Test Plan: Run petascope_insertdemo.sh. Check that the name of the single field of the coverage is Gray

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D89