ticket:1169 - Updated the maximum number of milliseconds allowed between two KeepAlive messages from the client.
Summary:
When executing long lasting queries, the server would fail with an exception. This was happening because the client was not sending KeepAlive messages often enough
to keep its "ALIVE" status on the server. Seeing that the client is not alive, the server would remove it from its registry.

Test Plan: Running long executing queries, including WCS and WCPS queries will not result in a timeout.

Reviewers: gmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D90