ticket:1185 - Added code in the new rasmgr to handle the old format of rasmgr.auth
Summary: Added code in rasmgr to support backwards compatibility with the previous format of rasmgr.auth

Test Plan: Manually tested with two different versions of the authentication file. At the moment, it is not possible to run unit tests because the code depends on the compile time variable CONFDIR. Unit tests are planned for a future release.

Reviewers: dmisev, mdumitru

Differential Revision: http://codereview.rasdaman.org/D105