ticket:1187 - Fixed bug in rascontrol grammar that did not allow peer host names to contain . . Fixed bug in rasmgr address binding
Summary: rasmgr was binding by default to localhost, this was changed to the default binding address for servers. A bug in the rascontrol grammar did not allow peer host names to contain ".".

Test Plan: Connecting to a rasmgr running inside a VM should work as expected if the VM has the necessary ports open.

Reviewers: gmerticariu, mdumitru, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D107