ticket:1171 - Cannot insert WMS layer if CRS definition is not available in opengis
Summary: When using wcst_import with wms:true to insert a WMS layer with CRS definition that does not exist in opengis, it will throw an exception and remove the layer. Then returns the coordinates of boudingbox in original CRS without transforming to WGS84.

Test Plan: using a custom definition CRS, such as http://localhost:8080/def/crs/PS/0/Mars to ingest coverage layer.

Reviewers: mdumitru, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D116