ticket:1192 - WCST_Import should catch WMSException when insert WMS layer error
Summary: In WCST_Import only catch WCSTException but not WMSException from WMSException.tpl.xml and returns success even when it has error in insert WMS layer. Also, it should not return stack trace to WCST_Import but write it to petascope.log.

Test Plan: Run WCST_Import.sh with a custom CRS definition not in opengis list and it will throw an WMSException but it is not logged in petascope.log also WCST_Impot does not handle it.

Reviewers: mdumitru, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D118