ticket:1252 - fix default tiling scheme
Summary: The default tiling scheme doesn't produce consistent tile sizes for 3D+ objects. This patch sets the default tiling scheme to [0:*,0:*,...] and lets the default tile size (4MB) determine the bounds of each tile, so that the result tiles are as close to 4MB as possible.

Test Plan: Updated systemtest

Reviewers: vmerticariu, mdumitru

Reviewed By: vmerticariu, mdumitru

Subscribers: pbaumann

Differential Revision: http://codereview.rasdaman.org/D145