ticket:1262 - Missing "+" sign when encoding from Petascope's console.
Summary:
This problem is from wcs-client, when only send access query to Petascope with parameter, starts with "&Request" and missing "service=WCS&version=2.0.1", then in PetascopeInterface
it will handle not correctly to the WCPS1.0 handler. Fix in wcs-client by adding WCS and version in SettingsService class.

Test Plan: Run the test case on WCPS test system.

Reviewers: mdumitru, atoader

Differential Revision: http://codereview.rasdaman.org/D163