ticket:1314 added parameter for disabling wcst in petascope.properties
Summary: Added "disable_write_operations" param in petascope.properties. When this is true, WCST profile does not appear in the capabilitis doc, and WCST requests are not processed. This is needed by people running public rasdaman servers who don't want to allow access to writes.

Test Plan: Deploy petascope, try to execute an insertCoverage requests (does not have to be valid, can be something like ?service=WCS&version=2.0.1&request=InsertCoverage which should result in a WCST specific error). Then set "disable_write_operations" to true, restart tomcat and retry the insert, which in this case should result in a parsing error.

Reviewers: mdumitru, dmisev

Differential Revision: http://codereview.rasdaman.org/D177