ticket:1188 - Make WCPS 1.5 ready
Summary: This patch contains the factoring of WCPS 1.5 for the purpose of using WCPS 1.5 instead of WCPS 1.0, it will also changes in WCS to use WCPS 1.5 for generating Rasql query instead of using XML from WCS then translate to Rasql as before. The most advatange of new WCPS 1.5 is it support better parsing error and is not needed to translate to XML then can generate Rasql query.

Test Plan: Run the test systems.

Differential Revision: http://codereview.rasdaman.org/D195