ticket:1350 - Petascope_WCS generate crsTransform invalid axis name
Before it is depend on the subset of WCS requests (e.g: &subset=Lat(0,20)&subset=Long(0,20)) to set the outputCrs in the crsTransform() of WCPS, e.g: crsTransform(c, {Lat:"http://.../4326"(0:20), Long:"http://..../4326"(0:20)

This is not correct if in WCS it don't have the subset parameters for the Lat, Long axis (or geo-referenced axes) then it should depend on the coverage metadata.

Test Plan: Run the test systems.

Differential Revision: http://codereview.rasdaman.org/D197