ticket:1301 - Plus sign is lost in WMS rasql fragment
The problem is when the request to Petascope has "+" which is not encoded and the HttpRequest will translate it to "" and it will be missing (e.g: in WMS Fragment: &rasqlFragment=case $iterator + 5 then {0,0,0} else {255,0,0} end will be translated to
&rasqlFragment=case $iterator 5 then {0,0,0} else {255,0,0}).

This will need to be fixed in Petascope to translate the queryString which is not encoded to the encoded correctly by a simple wrapper.

I have checked as well in test services, it works with request in KVP or XML from WCS or WCPS, the main test is in WMS with the purpose of this ticket.

Differential Revision: http://codereview.rasdaman.org/D201