ticket:1306 - Delete coverage by WCS will delete existing WMS layer if any
Summary: when remove coverage by WCS, it did not remove the WMS layers from this coverage and it is not consistent. Then, it needs to check if coverage has an exisiting WMS layer and other contents of this layer. All will need to be deleted.

Test Plan: The waxlake coverage is imported and then deleted by WCS when test wcst_import. A WMS test will try to delete this layerID again to check its existence.

Reviewers: mdumitru, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D218