ticket:1359 - fix connections in WMS
Summary: there is a mistake in WMS for opening connection without close it then it will occupy the connections each time has a new request (which is not cached), changed to use Interface instead of concrete object according to ormlite's example.

Test Plan: run the test_wms connections which will query multiple different requests and the connections should be stable (i.e: release connection as soon as possible).

Reviewers: mdumitru, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D219