ticket:1263 - Show nilValues in DescribeCoverage
Summary: The problem is from ps_quantity tables which the select query to check an existing quantity is not correct then it will not add new valid quantity with other null values. This is fixed to ingest new quantity when it is needed to this table.

Test Plan: test_time3d now is imported with the nullValues and in test_wcs will check the null values of this coverage when DescribeCoverage.

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D221