ticket:1387 added translation to origin in grib message domains
Summary: Message domains of grib files were previously relative to the domain of the object, not the file. As a consequence, the grib decoder was trying to read information from the wrong message # in some cases (where objects were formed of more grib files).

Test Plan: create an object from 2 or more grib files, try to read a region from the 2nd file.

Reviewers: dmisev, mdumitru

Differential Revision: http://codereview.rasdaman.org/D224