ticket:1400 - remove axes which contain axis iterator when slicing
Summary: In WCPS, subset contains AxisIterator, e.g: $px, $py will not be translated to numeric interval for geobounds/gridbounds, then when slice with axisIterator, it will still keep the original intervals, then will not strip it as the axis is not SlicingSubset. Then it needs to be stripped with the list of axis contains $.

Test Plan: run the test WCPS test.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D231