ticket:1410 - Fix scale, extend in WCPS, WCS
Summary: There was a misunderstanding in WCPS with scale() and extend(), e.g: scale(c, {Lat:"CRS:1"(0:300), Long:"CRS:1"(0:500)}) then outputCrs of c is CRS:1 which is not correct. Also fix scalefactor in WCS, it should scale down if scalefactor < 1 or scaleup if scalefactor > 1.

Test Plan: Run the test in test wcps, wcs

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D239