ticket:1426 added support for interval updates on irregular axes
Summary:
In order to allow this kind of import I had to 1) add functionality to
petascope to save multiple coefficients in a single request and 2) allow specifying these coefficients in wcst_import (as list of direct positions of the slices). Since this required changing the axis interface, I decided
to create a new recipe in order for all the previous ingredients to keep
working. I also tackled a different problem with the new recipe: in netcdf,
pixels are 0 dimensional and exist as a point in the middle of the interval.
Having a 1d object in netcdf, with 1 point and resolution 1, would translate
to hacing a point at coordinate 0.5. In petascope the same dataset translates
to having a point that stretches between 0 and 1. In order to accomodate this,
the minimums and the maximums are automatically adjusted by half resolution in
the recipe.

How ingredients look with the new recipe (general_cf_coverage):
"Lat": {
  "min": "${netcdf:variable:Lat:min}",
  "max": "${netcdf:variable:Lat:max}",
  "resolution": 10,
...

How ingredients look like with the old recipe (general_coverage), same dataset:

"Lat": {
  "min": "${netcdf:variable:Lat:min} - 5",
  "max": "${netcdf:variable:Long:max" + 5",
  "resolution": 10,
...

Both are available, and the old ones still work. I named the recipe "cf", as
all CF-compliant netCDFs consider pixels as 0 dim.

P.S.: @Bang our netcdf output doesn't consider pixels in the middle,
in order to be CF-compliant we need to shift them by half resolution.

Test Plan: Added systemtest

Reviewers: dmisev, bphamhuu

Subscribers: mdumitru

Differential Revision: http://codereview.rasdaman.org/D254