ticket:1425 - SECORE performance enhancement
Summary: The problem with SECORE when Tomcat restarted it will take ~1' to open the BaseX dbs. So this process need to be in initiliaziton process. Other important problem is the BaseX query will returns lot of unncessary datas and using Java to remove duplicates, sort data is expensive resources which will have problem with memory when have multiple requests.

Test Plan: Run the test with SECORE, before with 20 queries, SECORE will remove userdb when restart Tomcat or run out of memory, now it can pass all of these problems (restart Tomcat will not remove userdb and as used cache for query so other requests can be return in short time).

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D255