ticket:1442 - Fix error with netCDF when time axis is set "dataBound=false"
Summary: The problem is due to when update coverage (e.g: a 3D coverage and each update slice is 2D and time axis is not belonged to the slice (it is set "dataBound=false" in general coverage for netCDF), the time axis is still taken into account as a compoundCRS and that will overwrite the domainElements of inputCoverage's metadata (slice).

Test Plan: Import the data from MERIS and test system test for wcst_import.

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D267