ticket:1432 - clear Petascope, SECORE dependencies directory and build javadoc for SECORE without error.
Summary: As Petascope, SECORE used maven to build so the libraries (/lib) are not needed anymore, clean the pom.xml with mavan dependencies instead of local jar files and build javadoc for SECORE without error. Also when make install with cmake it will not build rasj, petascope, secore again.

Test Plan: tested with build with autotools and cmake.

Differential Revision: http://codereview.rasdaman.org/D282