ticket:1455 - wcst import add nodata values from image if they are not specified in ingredient files
Summary: There is a bug with gdal in wcst_import to read nodatavalue as they will be float number (e.g: 111 -> 111.0) so it must be parsed correctly. It will need to be fixed wrt to rasql problem in ticket 1464 as well.

Test Plan: eobstest has nodata_values then wcps, wcs encoding is sufficient.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D290