ticket:1377 - fix error with quote in CRS axis-label parameter
Summary: It seems that Petascope and SECORE in Debian both have error with quote as CRS parameter (e.g: AnsiDate?axis-label="t") which will need to be encoded to %22 (quote). Or it will fail when testing wcst_import for these cases.

Test Plan: it was tested directly in Debian VM and it passed all.

Reviewers: dmisev

Differential Revision: http://codereview.rasdaman.org/D292