ticket:1436 - WCPS encode netCDF with customizable range name
Summary: Normally, we don't need to pass the range name in encoding as it does not have any effect. However, in case of netcdf it can be taken into account as variable metadata. Then, the output will have the range which used defined in the WCPS query.

Test Plan: run the WCPS queries

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D304