ticket:1483 - fix memory leak occuring during ingestion with the Java API
Summary: Ingesting data with rasj creates a memory leak in rasserver, this patch fixes the issue.

Test Plan: It's tricky to add a test as the leak cannot be identified with valgrind for example. The Java code attached to the ticket can be used to manually observe if a leak happens by monitoring the memory usage of rasserver. With this patch the memory usage stays flat, so there is no memory leak.

Reviewers: vmerticariu

Differential Revision: http://codereview.rasdaman.org/D319