ticket:1485 - WCST_Import using decimal with precision
Summary:
This ticket tried to fix some problems:
+ Ingest coefficients from AnsiDate instead of UnixTime and add the option to select the different of AnsiDate origin (e.g: 1970-01-01 instead of 1601-01-01).
+ Get the values from netCDF variable without losing precision (by default, numpy array only returns float value and loose all the precisions, then must convert to string and parse).
+ Eval() also convert decimal to float by default, so need to reduce the cases which need to be evaluated to keep the precision.

Then, now, no need to add operation (*24*3600) in the ingredient file to convert date to seconds for Ansidate Date as it will return wrong date after the fix. And the resolution or values of netCDF variables will not lose the precision.

Test Plan: Run the test wcst_import

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D323