ticket:1475 - primary bugfix in ops.cc:5069. initVal was previously (mostly) uninitialized, leading to funny errors for structs. Notes on changes for ticket 1475 are as follows...
Summary:
(1) the destructor for NullValuesHandler was not virtual, but it is overridden in the code.

(2) spelling of "tupel" and "tuple" varied from file to file, so sometimes the variable was not the one we thought.
    I standardized its spelling to "tuple" from "tupel" everywhere, and fixed a few spelling mistakes in the documentation.

(3) changed myInitVal in GenCondenseOp to boolean from int since it is only used to flag whether or not initVal[] should be deleted in the destructor or not.

(4) changed init vals for min operators in op.cc to INFINITY for more consistency and because the min function for char had an init val of 0, which did not make sense.

(5) line 346 of qtcondenseop.cc, added "NULL" to the constructor for newInitVal -- probably not necessary, but it was being done in GenCondenseOp anyway.

(6) in GenCondenseOp's constructor, added a case for struct where all the initVals are initialized to be 0.

(7) added tests in test_select for addition, multiplication, min, max, some_cells, and all_cells.

caveat: It could be that people start getting errors for condense with structs, now that it works to begin with, but in this case
    someone should modify how initVal is handled in the structural case (e.g. reference initVal[i] for the [i]th element of the structure)

Test Plan: run test.sh in test_select (new tests with the query in question, along with a few similar queries, were added under rgb_*_select.rasql).

Reviewers: dmisev

Reviewed By: dmisev

Differential Revision: http://codereview.rasdaman.org/D325