ticket:1485 - Fix WCST_Import with subset AnsiDate when pixelIsPoint and enable subset_correction:true for TimeAxis
Summary:
This patch focused on fixing problem with netCDF with/without pixelIsPoint ingredient file. It also solved some small error for GRIB but there must be another patch to solve GRIB problems if any with/without pixelIsPoint soon.
Due to the change from float to Decimal in WCST_Import, the values from ncdump seems good to be pixelIsPoint (e.g: 75.05 75.10 75.15...) could be
(75.050000000003, 75.10000000007, 75.15000000000006) and the calculation in Petascope CrsComputer will not be as before (i.e: ingestable)
 so the usage of pixelIsPoint is reduced significantly (it is better to try without pixelIsPoint option first).

Test Plan: Run the system test for test_wcst_import

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D341