ticket:1485 - fix problem with wcst_import from float to decimal for gdal, netcdf recipes
Summary:
The patch tried to make the consistency between wcst_import and petascope, especially in how the total number of grid points is calculated.
So wcst_import needs to follow the formulation of petascope (math.ceil()/math.floor()) instead of round() as  before.
Because of evaluation netcdf, gribb variables, directions with eval(), they need to be converted to string, so ncdump shows value is: 179.05 but eval see it is 179.000000000005 as in string.

So to solve this tricky problem, an acceptable epsilon value (0.01) is added to avoid the increase/decrease of lower/upper grid point in both wcst_import and petascope.
For example: 4.99999 is considered as 5 and 5.00000001 is considered as 5, 4.8 or 5.1 is still the same their values.

For irregular time axis, changed to use BigDecimal as all the calculation was in BigDecimal already. But to support backward values in Double, a nother acceptable epsilon is: 0.00000000001 was added.
For example: coefficient in database is 0.0 (in double) and with BigDecimal it is 0.0000000000000000000000001 then consider it is 0.0.

This patch also tested regression (i.e: using before this patch for wcst_import, petascope insert coverages then use the patch and reingest into same coverage ids), here is the result:
https://docs.google.com/spreadsheets/d/1eOcp6t2YM1o8oESzQkLSeYyd647tiQygoe7LvjxiY-U/edit?usp=sharing
And they could be imported without error.

Test Plan: Run the test for wcst_import and tested for some collections from partners.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D347