ticket:1518 - Evaluation expression for metadata enquoted with single quotes by default
Summary:
All the metadata should be evaluated from global metadata or band's metadata. Also because they are string so must be enquouted with single quoutes in ingredient file.
It seems like this syntax only applied to "'${....}'" expression when user defined the ingredient files, but not with a string in metadata "This is a string", so the single quote will be added internally in this case for backwards compatibility.
NOTE: expression with multiple single quoutes like: '\\'\\'\\'\\'\\'\\'\\'a\\'\\'\\'\\'\\'\\'\\'\\'\\'' will return "a" by eval.
Also renamed test_data to testdata for consistence with other test folders.

Test Plan: Run the test for wcst_import

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D369