ticket:1524 - print feedback about which file is analyzing before importing by wcst_import
Summary:
wcst_import will analyze all the files to create the slices with the subsets, metadata,... before importing to Petascope.
It could take long time when the file is big or the number of files is large, so print what it is doing to know it still works.

Test Plan: Run the test wcst_import.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D370