ticket:1525 - Sort all the slices by the datetime axis in wcst_import
Summary:
Regular time series, Irregular time series already sorted slices by datetime, it needs to apply this function on the General recipes in cases the datetime are extracted from filenames,
but the list of files is not in order of datetime axis.

Test Plan: Run the test wcst_import with reversed file names.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D372