ticket:1528 - wcst_import keep the global attributes of netcdf when auto is specified in ingredient file
Summary:
If global:"auto" or global is not specified in ingredient file for netcdf, all the global attributes of netCDF file will be imported as global metadata.
Also fixed with eval to evaluate the metata correctly without problem for special characters.

Test Plan: Run the test for test_all_wcst_import

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D377