ticket:1536 - Support calculating netCDF regular axes's resolution internally in WCS_Import
Summary:
For netCdf file, it does not store the resolution variable as in GDAL and GRIB so user has to calculate it manually and it seems not correct when the coordinates for regular axis are not rounded.
So wcst_import will calculate the resolution based on all the coordinates when the expression is specified in ingredient file for netCdf.

Test Plan: Run the test wcst_import

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D380