ticket:1536 - petascope relaxes the difference between files's resolution
Summary: For netCDF files, the resolution can be different between files, such as for MERIS data: -0.002777777716032231175326479578 and -0.002777777719032231175326479578. So instead of checking it must have same resolution, it can be relaxed with a small epsilon when validating the input slice to update the target coverage.

Test Plan: It was tested to import MERIS coverage with different files when some files might have different resolution but it is small.

Differential Revision: http://codereview.rasdaman.org/D381