ticket:1540 - wcst_import with gdal error too many files open
Summary: gdal python limits the number of files could be opened to read dataset. In a test script from Dimitar, it can read to 1989 tiff files and get the error. So, it should only open file to get dataset for each iteration to extract metadata.

Test Plan: it is tested with Dimitar's cript and can start import coverages instead of error.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D382