ticket:1537 - Release memory from Petascope after finishing response
Summary:
Petascope need at least 1.5 * size of result from Rasj (ideally is 2) + 300MB for each WCS (WCPS) request to get a big result (hundred MBs). However, the gabarge collector does not release the byte[] when it finished the response, so need to do it manually.
Also the misleading error about the capabilitiy of servers is filtered more percisely.

Test Plan: It is tested with a WCPS request which returns ~600 MB and Petascope needs to set to 1200 MB to work. After each request, the heap cache is reduced to < 100 MB.

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D387