ticket:1552 - test if a string is json before attempting to parse it
Summary: Currently, we try to parse json; if this fails, we print a warning and try to parse key/value then. This is not the best approach, as sometimes the format parameters might be json, but fail to parse because the json is invalid. In this case rasdaman should throw an exception instead of printing a warning. A better test for whether the format parameters are json, is to check if the first non-whitespace character is '{'; this patch implements this solution.

Test Plan: Added a test to the conversion systemtest.

Reviewers: bbell, bphamhuu

Reviewed By: bphamhuu

Differential Revision: http://codereview.rasdaman.org/D395