ticket:1559 - WCPS: fix the fitToSampleSpace function correctly
Summary:
In WCPS, fitToSampleSpace function is used to fit the geo bounds (lower, upper) to a geo coordinate which corresponds to grid pixel.
It needs to be fixed correctly so the bounding box of WCPS encoding will return correct values for xmax, ymin, then the pixel size of output is also correct.

Test Plan: Run the test_all_wcst_import for the test coverage 10m and the output in WCPS result.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D397