ticket:1564 - WCST_Import InsertCoverage with first datetime as bound for time axis for old recipes
Summary:
The datetime when inserting coverage for time series regular, irregular is set 0 and this leads to wrong calculation in new Petascope for grid bound of time axis.
It should work as same as General Recipe when time axis also is added for InsertCoverage as first bound.

Test Plan: Tested with old Petascope and new Petascope.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D406