ticket:1574 - WCS-Client error with library when displaying a lot of coverages with GetCapabilities
Summary:
WCS-Client with the library code-pretty trying to parse the result from GetCapabilities in XML and put it in a text box with colorful text.
However, in case of PlanetServer-2 with very large XML result from GetCapabilities, this library cannot handle and Browser will freeze. It should only load a part of the content when it is too large to display in Browser.

Test Plan:
Tested with PlanetServer-2, takes 5 seconds to load 4500 coverages in web-client.
The maxium length is calculated with the need to display the whole DescribeCoverage result which is important.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D414