ticket:1565 - WCST_Import Send UpdateWcsLayer request to WMS handler instead of InsertWcsLayer request when layer already existed (new Petascope Part 1)
Summary:
WCS_Import did not check if a WMS layer with same name as coverage Id already existed in service endpoint so it always sends the request to import
a WMS layer with parameter InsertWCSLayer. It should send a request with UpdateWCSLayer when a WMS layer already imported
and update this layer with new metadata from update coverage.

Test Plan: Run test wcst_import for the test with duplicate wms_layer name.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D416