ticket:1569 - Reduce time for ingesting Grib file in wcst_import
Summary:
Grib file has problem with speed in wcst_import as it takes long time to open and iterate each message.
Then, need to remove all the unnecessary things like these.
The time to ingest a Grib file with 1612 messages (330 kbs) comes from: 26.725s to 13.271s.
In wcst_import, the number of messages is the most important value to consider,
for example: a much bigger file (160 MB with 310 messages take only ~2 seconds to analyze, but the file with 1612 messages will take 5 seconds to analyze).

Test Plan: run the test all wcst_import test

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D431