ticket:1593 - Supports bands's metadata and axes's metadata in WCST_Import, Petascope for encoding netCDF
Summary:
WCST_Import added the "bands" object as a property of "metadata" object in general recipes to allow to specify the metadata of
user defined coverage's bands. Then all these bands's metadata are collected as bands element of coverage's metadata in Petascope and
add it to each corresponding band of encoding in netCDF (only netCDF supports this feature).

Test Plan: Run the test for test_all_wcst_import and WCS, WCPS.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D435