ticket:1598 - Petascope supports delete coverage with error metadata
Summary:
DeleteCoverage requests needs to read the coverage's metadata before it can delete the Rasdaman collection and the coverage's metadata from petascopedb.
It cannot delete when the coverage's metadata has some errors and the coverage cannot be deleted in Petascope. It should only read the Rasdaman collection of the coverage to delete as this is the necessary metadata.

Test Plan: Tested with case when coverage's metadata error with irregular coefficitients.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D438