ticket:502 - Support to migrate to different datasource (DBMS) in new Petascope
Summary:
By default, Hibernate can only work with 1 datasource with Spring JPA. It has no problem to migrate from old petascopedb prior version 9.5 as
this old petascopedb used JDBC with native SQL, and new Petascope uses Hibernate with another connection to target datasource. However, in version 9.5, if one wants to migrate to different datasource (new database or new DBMS), this needs to configure Hibernate manually, so it can connect to both datasources.
Also, it reduced the time for querying basic metadata for WCS GetCapabilities when application starts as reading all the coverage's metadata is too long time and it is not needed.

Test Plan: Tested to migrate to new database in same DBMS, different DBMS (HSQL) and also in PlanetServer-2 (the total time to starts application with full cached basic metadata is ~3 minutes).

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D439