ticket:1490 - Insert WMS layer's style with wcpsQueryFragment besides rasqlTransformFragment
Summary:
WMS supports rasqlTransformFragment which is a kind of made-up query for ingesting style for a layer. Users can create
the query for a layer's style easier when using wcpsQueryFragment (the part of a normal WCPS query inside the encode operation).
Therefore, WCPS should be used instead of this rasqlTransform query, but it also keeps this feature for backwards compatibility.
NOTE: since Tomcat version 8, the characeters: '{', '}' are not valid for HTTP request, so must encode the query of rasqlTransformFragment or wcpsQueryFragment before sending to server in clients(e.g: WebBrowers).

Test Plan: Run the tests for test_wms.

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D445