ticket:502 - Fix logging in the Spring framework and faulty Bean injection in petascope
Summary:
There is a hidden problem from Spring Framework which it will stop Tomcat
then petascope also cannot start. The problem is shown when adding the log4j for Spring Framework when it starts and is fixed. It was about the ambiguity of injection Bean to a java class, but Spring cannot show the error in log
as Spring Framework starts before petascope.properties is loaded.

Test Plan: Tested by copying war, jar to a VM and it could run.

Reviewers: dmisev

Differential Revision: http://codereview.rasdaman.org/D457