ticket:1612 - Error when recreating duplicate cell/array/set types of reingesting coverage in Petascope
Summary:
After the patch for ticket 1029, the coverage base types (cell/array/set) is not created with random name anymore, but it uses the coverageId as the main part of the types's name.
Then, the check for existing base types in Petascope needs to be fixed strictly or it will try to recreate base types which are existing and throw exception.

Test Plan: Run the test for test wcst_import to ingest all the test coverages, then run the test clean up to remove all the coverages, then run the test for wcst_import again to reingest all coverages.

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D467