ticket:1602 - Fix WCPS translation in Petsacope from geo indices to grid indices correctly
Summary: The translation in Petascope for WCPS from geo indices to grid indices are not completely correct. Also, to make it correct, the most important value for coverage's axes are resolutions. They must be set almost correct from ingredient file to be imported by wcst_import, (e.g: the correct resolution is: 0.46666666666666666, then the input resolution should be: 0.46666666666666, not 0.466666).

Test Plan: It was validated from Oliver's test script in python: https://pastebin.com/0yegVrAx which will not return the diffecent between the python formula and WCPS imageCrsdomain for translating from geo indices to grid indices.

Reviewers: vmerticariu, dmisev

Differential Revision: http://codereview.rasdaman.org/D468